refactor(types): import EntityTypeString from the new dedicated module
